Move basic demo into its own component and explicitly install react-dnd
