+++ website/package.json
@@ -30,6 +30,8 @@
     "prism-react-renderer": "^1.3.5",
     "query-string": "^7.1.1",
     "react": "^18.2.0",
+    "react-dnd": "^16.0.1",
+    "react-dnd-html5-backend": "^16.0.1",
     "react-dom": "^18.2.0",
     "react-querybuilder": "next",
     "react-querybuilder-v4": "npm:react-querybuilder@^4.5.2",
