+++ website/src/pages/react-querybuilder/_QueryBuilderBootstrap.ts
@@ -1,0 +1,1 @@
+export { QueryBuilderBootstrap as default } from '@react-querybuilder/bootstrap';
